Uporządkowanie kolejności logiki i display - logika tylko w pliku Game.js
